requirements.txt
alembic==1.8.1
Babel==2.9.0
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
Flask==1.1.4
Flask-Migrate==3.1.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.2
importlib-metadata==2.1.3
importlib-resources==5.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.2.1
MarkupSafe==1.1.1
packaging==21.3
pathlib2==2.3.7.post1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.6.0
python-dotenv==0.20.0
pytz==2022.2.1
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.8.1
